06/requirements.txt
Flask==3.1.2
flask-restful
Flask-SQLAlchemy
flask-marshmallow
Flask-Migrate
marshmallow-sqlalchemy